With replay, check that spike time+delay is > t (resume time) before putting on the event queue.  This allows events that happen *close* to savetime can be reissued, but not break the queue.
Allow a new keyword/value into SpikeReplay stimulus.  'Timing Absolute' tells the simulator to not offset the times listed in a spike file by the simulation time (also ignores Delay keyword if included).  Useful for letting users reuse a spikefile on a resume sim without having to shift the timing of events to compensate for the save time.

Forgot to set Stimulus using TStims to be ignored by save/resume.  TStim's noise feature no longer puts a 0 as the first point of the voltage vector.  Noise stimulus now use the baseSeed argument to shift the random distributions.

For synapse models (ProbAMPANMDA_EMS.mod, ProbGABAAB_EMS.mod, ProbGABAA_EMS.mod) tsyn has been promoted to a template wide paramter rather than just a local to NET_RECEIVE block.  This is so that it can be included in the save/restore process.  Otherwise, on resume, it has a value of 0 which would be interpreted as the synapse not having been activated since time 0 and thus all synapses would likely fire at the next event.

JGK