As pointed out by Giuseppe, we can factor out some math from the synaptic integration process and compute these values once at the start of simulation.  Then apply these pre computed values directly.  This is detailed in JIRA ticket:
https://bbpteam.epfl.ch/project/issues/browse/BBPBGLIB-144

Testing show that a simulation of 31k cells for 4500ms went from needing 4197s to 3589s.  A 15% savings.  However some minor numerical epsilons seem to accumulate and membrane voltages and action potentials will start to shift slightly.  In the test this started to happen around 1258ms into the simulation.  If those are acceptable, then this branch can be merged into the master.

JGK