Latest changes for GapJunction functionality.
GapJunctions use the Projection block in the BLueConfig so that the typical chemical synapses
continue to use the old nrnPath of the Run block.

Also, update Reporting mod files to use different preprocessor directives such that the default behavior
is to enable things and one must pass an explicit flag to disable them.

JGK