Initial version to refactor synapse creation for ticket BBPBGLIB-227
Introduce the idea that *Helper.hoc objects handle the details of creating the synapse point processes.
If the users have a new custom synapse model which requires additional steps, then those are done in the
helper object.

In a BlueConfig, the override is given as an additional option 'ModOverride':

Connection synSwap
{
      Source Excitatory
 Destination post
      Weight 1
 ModOverride Demo2
}

In the above example, neurodamus would look for a template file Demo2Helper.hoc to load and instantiate.  That
helper then has the details for creating and setting values on some point process.

JGK