Update to StochKv to better handle switching to deterministic mode
When using ForwardSkip, dt is large, so StochKv will switch to deterministic mode.
Also, allow for either MCellRan4 or Random123 to be used as the rng.

Note that results will be different when ForwardSkip is used since the random stream does not pull values during the ForwardSkip time length.

https://bbpteam.epfl.ch/project/spaces/display/MEETBUILDSIM/StochKv+and+Random+Streams

StochKv issues that should be resolved with this check in:

General behavior due to large Dt
https://bbpteam.epfl.ch/project/issues/browse/BBPBGLIB-205

Proper warnings when values are out of range; not from simply using ForwardSkip
https://bbpteam.epfl.ch/project/issues/browse/BBPBGLIB-187
https://bbpteam.epfl.ch/project/issues/browse/BBPBGLIB-143

JGK

Change-Id: I0ff1c14959c0dcac1993852dac0bb7a4548b0a68