Additional changes to fix save/resume with latest NEURON
Connection.hoc

Need to ignore netstim_inhpoisson used by SpontMinis so they are not written to save file.

Node.hoc

Remove support for Replay.  Note that this is a short term solution since they will not be used for plasticity experiments.

Also ignore the BinReportHelper object itself

BinReportHelper.mod

Further refinement to timing in reports since previous commit where we had a conflict with loadbal computation.  So now with both use cases in mind, we make a distinction between activeStep and initialStep such that initialStep may be non-zero, but it is always set in the INITIAL block of the mod.

JGK