BBPBGLIB-972: mitigate endless loop
In cases like load-balance computation in NEURON, we might end up
calling this MOD file without setting up RNG. Previously this MOD
file was returning 0.5 from here and it was causing deadlock when
dt is higher (e.g. 0.1).

We looked at preivous versions of MOD files and they were returning
0. This avoid the deadlock when RNG is not setup.

As this MOD file is not supposed to be used without setting up RNG,
for the purpose of load-balancing and avoid deadlock, we are returning
0.