BBPBGLIB-972: Remove warnings from urand() to make it less verbose
In https://bbpgitlab.epfl.ch/hpc/sim/models/neocortex/-/merge_requests/13 (9063bd44f05e829c8ce366501e05a45eb793618e) we changed the return value of `urand()` function in case the random number generator is not set up to `0.0`. In this commit we remove the `WARNING` being printed in that case to avoid extremely verbose logs in that case.